attach app insights to API requests
