+++ packages/extension/src/constants.ts
@@ -1,6 +1,5 @@
 import vscode from 'vscode';
 
-export const EXTENSION_ID = "stateful.marquee";
 export const DEFAULT_FONT_SIZE = 5;
 
 export const RETRY = 4 * 1000;
