+++ packages/utils/tests/extension.test.ts
@@ -1,5 +1,5 @@
 import vscode from 'vscode';
-import ExtensionManager, { activate } from '../src/extension';
+import ExtensionManager, { activate, getExtProps } from '../src/extension';
 
 const context = {
   globalState: {
@@ -272,4 +272,10 @@ test('should upgrade config from v2 to v3', () => {
   context.globalState.get = jest.fn().mockReturnValue({ bg: 42 });
   const exp = activate(context as any, { appendLine: jest.fn() } as any);
   expect(exp.marquee.disposable.configuration.background).toBe(42);
+});
+
+test('should get proper extension props', () => {
+  const { extversion, ...snap } = getExtProps();
+  expect(snap).toMatchSnapshot();
+  expect(typeof extversion).toBe('string');
 });
