+++ packages/widget-github/tests/utils.test.ts
@@ -2,6 +2,7 @@ import { fetchData } from '../src/utils';
 
 declare const window: {
   fetch: Function
+  marqueeUserProps: any
   marqueeBackendBaseUrl: string
 };
 
@@ -11,6 +12,7 @@ beforeEach(() => {
 });
 
 test('fetchData', async () => {
+  window.marqueeUserProps = JSON.stringify({ foo: 'bar' });
   window.marqueeBackendBaseUrl = 'http://marquee.com';
   (window.fetch as jest.Mock).mockResolvedValue({
     ok: 1,
